Improve ListComponent item visibility
Added onListItemVisible(int index) and onListItemNotVisible(int index) to ListComponent.
Users can override these methods when they want to react to the list items being visible or invisible.
Unlike onItemVisible() or onItemNotVisible(), the provided index pertains to the actual items and not the dividers, so that users can use the provided index as the index of the items.